fix: add setu/gid handling for external libfuse
Many packagers use the external libfuse for new upstream improvements
and to avoid redundancy. There is no good technical reason to keep the
priviallage handling to internal libfuse only, so let's fix that.

I am slightly concerned at not seeing some of the mount point permission
checks found in upstream libfuse fusermount in our current mount
implementation. This would be a bug shared with internal libfuse, and I
do not intend to reach that far now.