chore: update kernel headers for upcoming change
The kernel headers are synced from upstream. I picked some libfuse3
changes in do_init as well, although in the spirit of our tiny private
library I greatly simplified the whole CAP thing so it becomes an
exact mapping of our kernel options.

The version negotiation is also greatly simplified. There is also no
need of not including ACL code in libfuse when ACL is off since the
capacity always guards against it. So yes, that one is gone too.